LayoutTests/fast/dom/Window/window-postmessage-clone-expected.txt

CONSOLE MESSAGE: line 54: RangeError: Maximum call stack size exceeded.
Tests that we clone object hierarchies
PASS: 'postMessage(cyclicObject)' threw ReferenceError: Can't find variable: cyclicObject
PASS: 'postMessage(cyclicObject)' threw TypeError: Cannot post cyclic structures.
PASS: 'postMessage(cyclicArray)' threw TypeError: Cannot post cyclic structures.
PASS: 'postMessage(reallyDeepArray)' threw RangeError: Maximum call stack size exceeded.
PASS: 'postMessage(window)' threw TypeError: Cannot post cyclic structures.
PASS: eventData is null of type object
PASS: eventData is undefined of type undefined
PASS: eventData is 1 of type number
PASS: eventData is true of type boolean
PASS: eventData is 1 of type string
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is of type object
PASS: eventData is 1,2,3 of type object
PASS: eventData is ,,1 of type object
PASS: eventData is null of type object
PASS: eventData is Wed Dec 31 1969 16:00:00 GMT-0800 (PST) of type object
PASS: eventData is of type object
PASS: eventData is done of type string
